add validate id

+++ backend/src/prices/prices.controller.ts
@@ -3,6 +3,7 @@ import {
   Controller,
   Delete,
   Get,
+  NotFoundException,
   Param,
   Post,
   Put,
@@ -17,6 +18,8 @@ import { PricesDto } from './price.dto';
 import { Price } from 'src/mongo/schemas/price.schema';
 import { Types } from 'mongoose';
 import { RoleGuard } from 'src/guards/roleGuard';
+import { Helpers } from 'src/projects/positions/helpers';
+import { ErrorsApp } from 'src/common/errors';
 
 @Controller('prices')
 export class PricesController {
@@ -46,12 +49,18 @@ export class PricesController {
     @Body() priceDto: PricesDto,
     @Req() req: RequestWithUser,
   ) {
+    if (!Helpers.checkId(priceId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(priceId);
     return this.pricesService.update(objectId, priceDto, req);
   }
 
   @Delete(':priceId')
   async remove(@Param('priceId') priceId: string, @Req() req: RequestWithUser) {
+    if (!Helpers.checkId(priceId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(priceId);
     return this.pricesService.remove(objectId, req);
   }
