+++ backend/src/projects/advances/advances.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  NotFoundException,
   Param,
   Patch,
   Post,
@@ -13,6 +14,8 @@ import { AdvancesService } from './advances.service';
 import { ProjectGuard } from '../project/project.guard';
 import { AdvanceDto } from './advance.dto';
 import { Types } from 'mongoose';
+import { Helpers } from '../positions/helpers';
+import { ErrorsApp } from 'src/common/errors';
 
 @Controller('advances')
 export class AdvancesController {
@@ -22,6 +25,9 @@ export class AdvancesController {
   @UsePipes(new ValidationPipe())
   @UseGuards(ProjectGuard)
   async create(@Body() dto: AdvanceDto, @Param('projectId') projectId: string) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.advancesService.createAdvances(dto, objectId);
   }
@@ -34,6 +40,9 @@ export class AdvancesController {
     @Param('projectId') projectId: string,
     @Param('advancesId') advancesId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     return await this.advancesService.updateAdvance(dto, objectId, advancesId);
   }
@@ -43,6 +52,9 @@ export class AdvancesController {
     @Param('projectId') projectId: string,
     @Param('advancesId') advancesId: string,
   ) {
+    if (!Helpers.checkId(projectId)) {
+      throw new NotFoundException(ErrorsApp.BED_ID);
+    }
     const objectId = new Types.ObjectId(projectId);
     await this.advancesService.removeAdvance(objectId, advancesId);
   }
